Add commands and tests
